refactor(feature-info): move files to corresponding locations
