+++ src/components/info/templates/default-template.vue
@@ -1,6 +1,9 @@
 <script setup lang="ts">
 import { defineProps } from 'vue'
-import { FeatureInfoJSON, FeatureJSON } from '../feature-info.model'
+import {
+  FeatureInfoJSON,
+  FeatureJSON,
+} from '../../../services/info/feature-info.model'
 import { useTranslation } from 'i18next-vue'
 import {
   hasAttributes,
